requirements_dev.txt
# Development, use pip install -e .[dev]
coverage
black
mypy
proxy.py
pytest
pytest-cov
pytest-mypy
Sphinx>=2.4.0
sphinx-rtd-theme
twine
pytest-flake8
isort
darglint
requests-mock